refac: 피드백 반영 6 - PlayingCards가 Cards를 상속하면서 조합하던 것을 조합만 하게 수정
