+++ src/main/java/domain/card/PlayingCards.java
@@ -1,21 +1,22 @@
 package domain.card;
 
+import sun.plugin.dom.exception.InvalidStateException;
+
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
-public class PlayingCards extends Cards {
+public class PlayingCards {
     private static final int BLACK_JACK = 21;
     private static final int MIN_SIZE = 2;
 
     private final Cards cards;
 
     private PlayingCards(List<Card> cards) {
-        super(cards);
         this.cards = Cards.of(cards);
     }
 
     private PlayingCards(Cards cards) {
-        super(cards);
         this.cards = cards;
     }
 
@@ -50,6 +51,37 @@ public class PlayingCards extends Cards {
 
     public List<String> serialize() {
         return cards.serialize();
+    }
+
+    public int size() {
+        return cards.size();
+    }
+
+    private int calculateSumWithAces(int sum) {
+        if (cards.hasAce()) {
+            Cards aces = getAces();
+            return calculateSumWithAces(sum, aces);
+        }
+        return sum;
+    }
+
+    private int calculateSumWithAces(int sum, Cards aces) {
+        if (aces.hasCardNotAce()) {
+            throw new InvalidStateException("복수의 에이스 카드 내에 부적절한 카드가 존해합니다.");
+        }
+        for (Card card : aces.getCards()) {
+            int score = card.calculate(sum);
+            sum += score;
+        }
+        return sum;
+    }
+
+    private Cards getAces() {
+        List<Card> aces = this.cards.getCards().stream().filter(Card::isAce).collect(Collectors.toList());
+        if (aces.isEmpty()) {
+            throw new InvalidStateException(String.format("%s가 존재하지 않습니다.", Symbol.ACE.getPattern()));
+        }
+        return Cards.of(aces);
     }
 
     @Override
